[feat/meongle-2] Add Common Response, Global Exception
