+++ src/main/java/ddog/mungleserver/global/auth/config/enums/Provider.java
@@ -1,0 +1,5 @@
+package ddog.mungleserver.global.auth.config.enums;
+
+public enum Provider {
+    KAKAO
+}
