+++ src/main/java/ddog/mungleserver/global/auth/config/enums/Role.java
@@ -1,0 +1,14 @@
+package ddog.mungleserver.global.auth.config.enums;
+
+public enum Role {
+
+    CUSTOMER("사용자"),
+    GROOMER("미용사"),
+    ADMIN("관리자");
+
+    private String description;
+
+    Role(String description) {
+        this.description = description;
+    }
+}
