+++ src/main/java/ddog/mungleserver/global/auth/config/jwt/JwtAuthenticationEntryPoint.java
@@ -1,0 +1,19 @@
+package ddog.mungleserver.global.auth.config.jwt;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+    @Override
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
+        /* 유효한 자격증명을 제공하지 않고 접근할 시 401 에러 */
+        response.sendError(HttpServletResponse.SC_UNAUTHORIZED);
+    }
+}
