+++ src/main/java/ddog/mungleserver/global/auth/dto/KakaoAccessTokenDto.java
@@ -1,0 +1,12 @@
+package ddog.mungleserver.global.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class KakaoAccessTokenDto {
+    private String kakaoAccessToken;
+}
