+++ src/main/java/ddog/mungleserver/global/auth/dto/RefreshTokenDto.java
@@ -1,0 +1,8 @@
+package ddog.mungleserver.global.auth.dto;
+
+import lombok.Getter;
+
+@Getter
+public class RefreshTokenDto {
+    private String refreshToken;
+}
