+++ src/main/java/ddog/mungleserver/global/auth/dto/TokenInfoDto.java
@@ -1,0 +1,15 @@
+package ddog.mungleserver.global.auth.dto;
+
+import ddog.mungleserver.global.auth.config.enums.Role;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class TokenInfoDto {
+
+    private final String grantType;
+    private final String accessToken;
+    private final String refreshToken;
+    private final Role role;
+}
