+++ src/main/java/ddog/mungleserver/global/exception/BadRequestException.java
@@ -1,0 +1,7 @@
+package ddog.mungleserver.global.exception;
+
+public class BadRequestException extends CustomRuntimeException {
+    public BadRequestException(String message, Object... args) {
+        super(message, args);
+    }
+}
