+++ src/main/java/ddog/mungleserver/global/exception/CustomRuntimeException.java
@@ -1,0 +1,9 @@
+package ddog.mungleserver.global.exception;
+
+import org.apache.logging.log4j.message.ParameterizedMessage;
+
+public class CustomRuntimeException extends RuntimeException {
+    public CustomRuntimeException(String message, Object... args) {
+        super(ParameterizedMessage.format(message, args));
+    }
+}
