+++ src/main/java/ddog/mungleserver/global/exception/IllegalParameterException.java
@@ -1,0 +1,7 @@
+package ddog.mungleserver.global.exception;
+
+public class IllegalParameterException extends RuntimeException {
+    public IllegalParameterException(String message, Object... args) {
+        super();
+    }
+}
