+++ src/main/java/ddog/mungleserver/global/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package ddog.mungleserver.global.exception;
+
+public class NotFoundException extends CustomRuntimeException {
+    public NotFoundException(String message, Object... args) {
+        super(message, args);
+    }
+}
