--- src/main/java/ddog/mungleserver/infrastructure/auth/config/enums/Provider.java
@@ -1,5 +1,0 @@
-package ddog.mungleserver.infrastructure.auth.config.enums;
-
-public enum Provider {
-    KAKAO
-}
