--- src/main/java/ddog/mungleserver/infrastructure/auth/config/jwt/JwtAccessDeniedHandler.java
@@ -1,19 +1,0 @@
-package ddog.mungleserver.infrastructure.auth.config.jwt;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        /* 권한 없는 곳에 접근할 경우 403 에러 */
-        response.sendError(HttpServletResponse.SC_FORBIDDEN, accessDeniedException.getMessage());
-    }
-}
