--- src/main/java/ddog/mungleserver/infrastructure/auth/config/jwt/JwtAuthenticationFilter.java
@@ -1,54 +1,0 @@
-package ddog.mungleserver.infrastructure.auth.config.jwt;
-
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletRequest;
-import jakarta.servlet.ServletResponse;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.GenericFilterBean;
-
-import java.io.IOException;
-
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends GenericFilterBean {
-
-    private final JwtTokenProvider jwtTokenProvider;
-
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-        String token = resolveToken((HttpServletRequest) request);
-
-        try {
-            if (token != null && jwtTokenProvider.validateToken(token)) {
-                Authentication authentication = jwtTokenProvider.getAuthentication(token);
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
-        } catch (ExpiredJwtException e) {
-            /* 추후에 EXPIRED_TOKEN 으로 변경 예정 */
-            throw new RuntimeException();
-        } catch (UnsupportedJwtException e) {
-            /* 추후에 UNSUPPORTED_TOKEN 으로 변경 예정 */
-            throw new RuntimeException();
-        } catch (Exception e) {
-            /* 추후에 UNKNOWN_ERROR 으로 변경 예정 */
-            throw new RuntimeException();
-        }
-
-        chain.doFilter(request, response);
-    }
-
-    private String resolveToken(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7);
-        }
-
-        return null;
-    }
-}
