--- src/main/java/ddog/mungleserver/infrastructure/auth/dto/KakaoAccessTokenDto.java
@@ -1,12 +1,0 @@
-package ddog.mungleserver.infrastructure.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class KakaoAccessTokenDto {
-    private String kakaoAccessToken;
-}
