--- src/main/java/ddog/mungleserver/infrastructure/auth/dto/RefreshTokenDto.java
@@ -1,8 +1,0 @@
-package ddog.mungleserver.infrastructure.auth.dto;
-
-import lombok.Getter;
-
-@Getter
-public class RefreshTokenDto {
-    private String refreshToken;
-}
