--- src/main/java/ddog/mungleserver/infrastructure/auth/dto/TokenInfoDto.java
@@ -1,15 +1,0 @@
-package ddog.mungleserver.infrastructure.auth.dto;
-
-import ddog.mungleserver.infrastructure.auth.config.enums.Role;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class TokenInfoDto {
-
-    private final String grantType;
-    private final String accessToken;
-    private final String refreshToken;
-    private final Role role;
-}
