+++ src/main/java/ddog/mungleserver/infrastructure/jpa/Customer.java
@@ -1,0 +1,30 @@
+package ddog.mungleserver.infrastructure.jpa;
+
+import ddog.mungleserver.global.auth.config.enums.Provider;
+import ddog.mungleserver.global.auth.config.enums.Role;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+public class Customer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String email;
+    private String nickname;
+
+    @Enumerated(value = EnumType.STRING)
+    private Provider provider;
+
+    @Enumerated(value = EnumType.STRING)
+    private Role role;
+}
