--- src/main/java/ddog/mungleserver/jpa/customer/Customer.java
@@ -1,30 +1,0 @@
-package ddog.mungleserver.jpa.customer;
-
-import ddog.mungleserver.infrastructure.auth.config.enums.Provider;
-import ddog.mungleserver.infrastructure.auth.config.enums.Role;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Customer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String email;
-    private String nickname;
-
-    @Enumerated(value = EnumType.STRING)
-    private Provider provider;
-
-    @Enumerated(value = EnumType.STRING)
-    private Role role;
-}
